Update to use pre-calculated boxes (geotherm) and recalc VBR only
Mostly, need to interpolate between the absolute T used in the
first half of the calculation and the potential T sweep used in
the pre-saved box.  Not entirely sure I've remembered to convert
at every stage from T to Tp etc.  Also, think best_T_g_phi is
not working properly.  AND need to make sure that melt is put
in following a step-function beneath zPlate ONLY.